# kvcoord: use `sync.Map` for DistSender circuit breakers